+++ CHANGELOG.md
@@ -26,6 +26,8 @@ The format of the release notes follows [Keep a Changelog](https://keepachangelo
 
 ### Fixed
 
+- converted file sizes are only kept in the database after an update, if the original file size
+  hasn't changed ([#20])
 - Automatically update version number in the badge in the README and the link to the latest stable
   during pre-release
 - Fixed an issue with the setup of the target extension that lead to errors during download sessions
@@ -166,3 +168,4 @@ prevent LibCurl error 9
 [v0.5.0]: https://github.com/LIM-AeroCloud/ICARE.jl/releases/tag/v0.5.0
 [v0.5.1]: https://github.com/LIM-AeroCloud/ICARE.jl/releases/tag/v0.5.1
 [#18]: https://github.com/LIM-AeroCloud/ICARE.jl/issues/18
+[#20]: https://github.com/LIM-AeroCloud/ICARE.jl/issues/20